ADD tests for contentAction
