+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^4.6.0",
+    "core-util-is": "^1.0.2",
     "html-react-parser": "^1.2.4",
     "jquery": "^3.5.1",
     "popper.js": "^1.16.1",
@@ -35,13 +36,13 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.20.6",
     "eslint-plugin-react-hooks": "^2.5.1",
-    "stylelint": "^13.3.3",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.17.2",
     "react-addons-test-utils": "^15.6.2",
     "react-test-renderer": "^17.0.1",
-    "stylelint-config-recommended": "^3.0.0"
+    "stylelint": "^13.3.3",
+    "stylelint-config-recommended": "^3.0.0",
+    "stylelint-config-standard": "^20.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.17.2"
   },
   "scripts": {
     "start": "react-scripts start",
