+++ src/actions/contentAction.js
@@ -19,6 +19,7 @@ export const fetchContent = (version, passage) => dispatch => {
   fetch(`https://api.biblia.com/v1/bible/content/${version}.html.json?passage=${passage}&style=oneVersePerLine&formatting=all&fullText=false&key=fd37d8f28e95d3be8cb4fbc37e15e18e`)
     .then(response => response.json())
     .then(data => {
+      console.log(data);
       dispatch(contentFetchSuccess(data));
       return data;
     })
